Chore: Refactor code
