+++ src/routes/user.routes.js
@@ -3,13 +3,13 @@ const express = require('express');
 const userController = require('../controllers/user.controller');
 const validateRequests = require('../middlewares/request.validator');
 // require schemas
-const schemas = require('../middlewares/schemas.validator');
+
 
 const router = express.Router();
 
 router.route('/users').get( validateRequests.reqAuthValidator,userController.getUsers)
-  .post( validateRequests.reqAuthValidator,userController.createUser);
-router.post('/users/register', userController.createUserLogin);
+                      .post( validateRequests.reqAuthValidator,userController.createUser);
+
 
 
 module.exports = router;
