--- src/services/user.service.js
@@ -1,11 +1,0 @@
-
-const { users } = require('../models');
-
-const userServices = {
-  getAllUsers: async () => {
-    const allUsers = await users.findAll();
-    return allUsers;
-  }
-};
-
-module.exports = userServices;
